Move mol2/pdb parsing methods to parser module
