+++ opencadd/databases/klifs/core.py
@@ -1253,66 +1253,3 @@ class CoordinatesProvider(BaseProvider):
         if output_format:
             if output_format == "rdkit" and entity != "ligand":
                 raise ValueError(f"Only entity ligand can be fetched as rdkit molecule.")
-
-    def _split_mol2_subst_names(self, mol2_df):
-        """
-        Split "residue.subst_name" column values from mol2 file and add as 
-        "residue.name" and "residue.pdb_id" columns.
-
-        Parameters
-        ----------
-        mol2_df : pandas.DataFrame
-            Structural data.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Structural data, including additional columns for residue name and PDB ID.
-        """
-
-        result = mol2_df.apply(
-            lambda x: self._split_mol2_subst_name(x["residue.subst_name"], x["atom.type"]), axis=1,
-        )
-        res_names = [res_name for res_name, res_id in result]
-        res_ids = [res_id for res_name, res_id in result]
-
-        mol2_df["residue.name"] = res_names
-        mol2_df["residue.pdb_id"] = res_ids
-
-        return mol2_df
-
-    @staticmethod
-    def _split_mol2_subst_name(subst_name, atom_type):
-        """
-        Split "residue.subst_name" values from mol2 file.
-
-        Parameters
-        ----------
-        subst_name : str
-            Substructure name in mol2 file. 
-        atom_type : str
-            Atom type.
-
-        Returns
-        -------
-        tuple (str, int)
-            Residue name and residue PDB ID.
-
-        Notes
-        -----
-        Regular example: "ALA1" ("residue.subst_name") will be split to 
-        - "ALA" ("residue.name") and 
-        - 1 ("residue.pdb_id").
-        """
-
-        # Handle "residues" that are elements such as CA or MG.
-        if subst_name[:2] == atom_type.upper():
-            res_name = subst_name[:2]
-            res_id = subst_name[2:]
-
-        # These are amino acid, linkers, compounds, ...
-        else:
-            res_name = subst_name[:3]
-            res_id = subst_name[3:]
-
-        return res_name, res_id
